refactor: MVC 패턴에 따른 패키지 분리
