+++ src/main/java/racingcar/domain/Car.java
@@ -1,6 +1,4 @@
-package racingcar;
-
-import java.util.Optional;
+package racingcar.domain;
 
 public class Car {
