+++ src/main/java/racingcar/domain/CarGame.java
@@ -1,6 +1,5 @@
-package racingcar;
+package racingcar.domain;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class CarGame {
