+++ src/main/java/racingcar/domain/RandomMovingStrategy.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class RandomMovingStrategy implements MovingStrategy {
