+++ src/test/java/racingcar/domain/CarGroupTest.java
@@ -1,7 +1,9 @@
-package racingcar;
+package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.CarGroup;
 
 import java.util.ArrayList;
 import java.util.Arrays;
