+++ src/test/java/racingcar/domain/CarNameTest.java
@@ -1,8 +1,9 @@
-package racingcar;
+package racingcar.domain;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.CarName;
 
 import static org.assertj.core.api.Assertions.assertThat;
