+++ src/test/java/racingcar/domain/CarPositionTest.java
@@ -1,8 +1,8 @@
-package racingcar;
+package racingcar.domain;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.CarPosition;
 
 import static org.assertj.core.api.Assertions.assertThat;
