+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,9 +1,9 @@
-package racingcar;
+package racingcar.domain;
 
 
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
