+++ src/test/java/racingcar/view/CarInputViewTest.java
@@ -1,7 +1,8 @@
-package racingcar;
+package racingcar.view;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.view.CarInputView;
 
 import java.util.List;
